refactor: move test utils to internal and rename
