+++ internal/test_utils/loaders.go
@@ -1,5 +1,5 @@
 // Package tests provides test types and functions for yac-p
-package tests
+package test_utils
 
 // GetTestConfigLoader returns a function that loads a static test configuration
 func GetTestConfigLoader() func() ([]byte, error) {
