+++ pkg/yace/yace_test.go
@@ -3,7 +3,7 @@ package yace
 import (
 	"testing"
 
-	"github.com/kjansson/yac-p/pkg/tests"
+	"github.com/kjansson/yac-p/internal/test_utils"
 	"github.com/kjansson/yac-p/pkg/types"
 )
 
@@ -11,7 +11,7 @@ func TestConfigLoad(t *testing.T) {
 	c := &YaceClient{}
 
 	conf := types.Config{
-		ConfigFileLoader: tests.GetTestConfigLoader(),
+		ConfigFileLoader: test_utils.GetTestConfigLoader(),
 	}
 
 	err := c.Init(conf) // Initialize all components
